refactor: streamline password reset flow by consolidating components
- Removed `OtpCodeComponent` and `ResetPassComponent` in favor of a single `ForgetPassComponent` that handles email submission, OTP verification, and password reset in a step-by-step manner.
- Updated routing to reflect the new structure and removed obsolete routes.
- Enhanced the UI for better user experience with improved styling and layout.
- Adjusted form handling and validation logic to accommodate the new flow.
