+++ apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/forget-pass/forget-pass.component.html
@@ -1,24 +1,53 @@
-<div>
-  <p class="text-2xl font-bold text-center m-2">Forgot Password</p>
-</div>
+<div class="forget-pass-container">
+  <!-- Step 1: Email -->
+  @if (currentStep === Step.Email) {
+    <div class="step-form">
+      <h2 class="text-center">Forgot Password</h2>
+      <form [formGroup]="emailForm" (ngSubmit)="submitEmail()">
+        <div class="input-group">
+          <input type="email" formControlName="email" placeholder="Email" aria-label="Email" />
+        </div>
+        <button type="submit" [disabled]="emailForm.invalid" class="btn">Send Reset Code</button>
+      </form>
+    </div>
+  }
 
-<div *ngIf="errorMessage" class="error-message">{{ errorMessage }}</div>
-<div *ngIf="successMessage" class="success-message">{{ successMessage }}</div>
+  <!-- Step 2: OTP -->
+  @if (currentStep === Step.Otp) {
+    <div class="step-form">
+      <h2 class="text-center">OTP Code</h2>
+      <p class="text-center">Enter the 6-digit code sent to your email</p>
+      <form [formGroup]="otpForm" (ngSubmit)="submitOtp()">
+        <div class="otp-inputs">
+          <input #otpInput *ngFor="let i of [1,2,3,4,5,6]" type="text" maxlength="1" [formControlName]="'digit' + i" (input)="moveToNext($event, i)" /> 
 
-<div class="form-box">
-  <form [formGroup]="forgetForm" (ngSubmit)="submit()">
-    <input
-      type="email"
-      placeholder="Email"
-      formControlName="email"
-      required
-    />
-    <button type="submit" class="btn" [disabled]="forgetForm.invalid">
-      Send Reset Code
-    </button>
-    <p class="switch mt-2">
-      Remembered your password?
-      <a routerLink="/auth/login">Login</a>
-    </p>
-  </form>
+         
+        </div>
+        <button type="submit" [disabled]="otpForm.invalid" class="btn">Confirm</button>
+      </form>
+    </div>
+  }
+
+  <!-- Step 3: Reset Password -->
+  @if (currentStep === Step.Reset) {
+    <div class="step-form">
+      <h2 class="text-center">Reset Password</h2>
+      <form [formGroup]="resetForm" (ngSubmit)="submitReset()">
+        <div class="input-group">
+          <input type="email" formControlName="email" placeholder="Email" readonly />
+        </div>
+        <div class="input-group">
+          <input type="password" formControlName="newPassword" placeholder="New Password" aria-label="New Password" />
+        </div>
+        <button type="submit" [disabled]="resetForm.invalid" class="btn">Reset Password</button>
+      </form>
+    </div>
+  }
+
+  <!-- Error/Success Messages -->
+  @if (errorMessage) {
+    <div class="error-message">
+      <span>⚠️</span> {{ errorMessage }}
+    </div>
+  }
 </div>
