+++ apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/forget-pass/forget-pass.component.scss
@@ -1,20 +1,32 @@
-.form-box {
+.forget-pass-container {
   margin: auto !important;
-  width: 80%;
-  border-radius: 50px;
-  border-width: 1px;
-  gap: 8px;
-  padding: 40px;
+  max-width: 400px;
+  width: 100%;
+  border-radius: 32px;
+  border: 1.5px solid rgba(255, 255, 255, 0.35);
+  background: rgba(30, 30, 30, 0.65);
+  box-shadow: 0 4px 32px rgba(0, 0, 0, 0.25);
+  padding: 48px 36px 36px 36px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 
-  form {
+  .step-form {
+    width: 100%;
+    display: flex;
     flex-direction: column;
-    gap: 0;
+    gap: 1.5rem;
   }
 
-  input {
+  .input-group {
     width: 100%;
-    margin-bottom: 14px;
-    padding: 8px 16px;
+    margin-bottom: 1rem;
+  }
+
+  input[type='email'],
+  input[type='password'] {
+    width: 100%;
+    padding: 0.75rem 1rem;
     border: 1.5px solid rgba(255, 255, 255, 0.18);
     border-radius: 20px;
     background: rgba(255, 255, 255, 0.10);
@@ -24,83 +36,43 @@
     font-weight: 400;
     transition: border 0.2s;
     box-shadow: none;
-    gap: 8px;
-
+    margin-bottom: 14px;
     &::placeholder {
       color: #bbb;
       opacity: 1;
     }
-
     &:focus {
       border: 1.5px solid #ff4100;
       background: rgba(255, 255, 255, 0.16);
     }
   }
 
-  .forgot {
-    text-align: right;
-    font-size: 13px;
-    margin-bottom: 18px;
-
-    a {
-      color: #ff4100;
-      text-decoration: none;
-      font-weight: 500;
-    }
-  }
-
-  .divider {
-    text-align: center;
-    margin: 18px 0 18px 0;
-    color: #bbb;
-    position: relative;
-    font-size: 15px;
-    font-weight: 500;
-
-    &::before,
-    &::after {
-      content: '';
-      position: absolute;
-      top: 50%;
-      width: 40%;
-      height: 1.5px;
-      background: #bbb;
-    }
-
-    &::before {
-      left: 0;
-    }
-
-    &::after {
-      right: 0;
-    }
-  }
-
-  .social-icons {
+  .otp-inputs {
     display: flex;
     justify-content: center;
-    gap: 18px;
-    margin-bottom: 20px;
+    gap: 14px;
+    margin-bottom: 24px;
+    input {
+      @extend .otp-box;
+    }
+  }
 
-    i {
-      color: rgb(255, 255, 255);
-      padding: 8px;
-      border-radius: 50%;
-      font-size: 16px;
-      width: 38px;
-      height: 38px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      cursor: pointer;
-      transition: transform 0.3s;
-      background-color: black;
-
-      &:hover {
-        transform: scale(1.1);
-        background: #ff4100;
-        color: #fff;
-      }
+  .otp-box {
+    width: 40px;
+    height: 48px;
+    font-size: 2rem;
+    text-align: center;
+    border: none;
+    border-bottom: 3px solid #fff;
+    background: transparent;
+    color: #ff4100;
+    font-weight: 700;
+    outline: none;
+    margin: 0 2px;
+    border-radius: 0;
+    transition: border-color 0.2s;
+    &:focus {
+      border-bottom: 3px solid #ff4100;
     }
   }
 
@@ -109,7 +81,7 @@
     background: #ff4100;
     padding: 15px;
     border: none;
-    border-radius: 12px;
+    border-radius: 24px;
     color: white;
     font-weight: bold;
     font-size: 18px;
@@ -117,21 +89,8 @@
     transition: background 0.3s;
     margin-top: 8px;
     margin-bottom: 0;
-
     &:hover {
       background: #e83e00;
-    }
-  }
-
-  .switch {
-    margin-top: 18px;
-    font-size: 15px;
-    text-align: center;
-
-    a {
-      color: #ff4100;
-      text-decoration: none;
-      font-weight: 500;
     }
   }
 }
@@ -150,16 +109,14 @@
   letter-spacing: 0.5px;
 }
 
-.success-message {
-  background: #22c55e;
-  color: #fff;
-  padding: 12px 20px;
-  border-radius: 10px;
-  margin: 16px auto 0 auto;
-  text-align: center;
-  font-size: 16px;
-  font-weight: 600;
-  max-width: 80%;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  letter-spacing: 0.5px;
+
+@media (max-width: 500px) {
+  .forget-pass-container {
+    padding: 24px 8px 16px 8px;
+    max-width: 98vw;
+    .otp-box {
+      width: 32px;
+      font-size: 1.3rem;
+    }
+  }
 }
