+++ apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/forget-pass/forget-pass.component.ts
@@ -1,46 +1,115 @@
-import { Component, inject } from '@angular/core';
+import { Component, ElementRef, QueryList, ViewChildren, inject } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
 import { AuthApiService } from 'apps/SuperFitnessApp/src/lib/auth-api/src/public-api';
 import { Subject, takeUntil } from 'rxjs';
-import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
+import {   Router } from '@angular/router';
+
+enum Step {
+  Email = 'Email',
+  Otp = 'Otp',
+  Reset = 'Reset',
+}
 
 @Component({
   selector: 'app-forget-pass',
   standalone: true,
-  imports: [CommonModule, ReactiveFormsModule],
+  imports: [CommonModule, ReactiveFormsModule, ],
   templateUrl: './forget-pass.component.html',
-  styleUrls: ['./forget-pass.component.scss']
+  styleUrls: ['./forget-pass.component.scss'],
 })
 export class ForgetPassComponent {
   private ngUnsubscribe = new Subject<void>();
   private _authApiService = inject(AuthApiService);
   private fb = inject(FormBuilder);
-  private _router = inject(Router);
+  private router = inject(Router);
+  Step = Step;
+  currentStep: Step = Step.Email;
 
-  forgetForm: FormGroup = this.fb.group({
+  emailForm = this.fb.group({
     email: ['', [Validators.required, Validators.email]]
   });
+
+  otpForm: FormGroup = this.fb.group({
+    digit1: ['', [Validators.required, Validators.pattern(/^[0-9]$/)]],
+    digit2: ['', [Validators.required, Validators.pattern(/^[0-9]$/)]],
+    digit3: ['', [Validators.required, Validators.pattern(/^[0-9]$/)]],
+    digit4: ['', [Validators.required, Validators.pattern(/^[0-9]$/)]],
+    digit5: ['', [Validators.required, Validators.pattern(/^[0-9]$/)]],
+    digit6: ['', [Validators.required, Validators.pattern(/^[0-9]$/)]]
+  });
+
+  resetForm = this.fb.group({
+    email: ['', [Validators.required, Validators.email]],
+    newPassword: ['', [Validators.required, Validators.minLength(8)]]
+  });
+
+  @ViewChildren('otpInput') otpInputs!: QueryList<ElementRef>;
 
   errorMessage = '';
   successMessage = '';
 
-  submit() {
-    if (this.forgetForm.invalid) return;
-    this._authApiService.Forgetpass(this.forgetForm.value)
+  submitEmail() {
+    if (this.emailForm.invalid) return;
+    this._authApiService.Forgetpass({ email: this.emailForm.get('email')?.value || '' })
       .pipe(takeUntil(this.ngUnsubscribe))
       .subscribe({
         next: () => {
-          this.successMessage = 'Check your email for the verification code.';
           this.errorMessage = '';
-          // Optionally navigate to OTP page
-          this._router.navigate(['/auth/otp-code']);
+          this.currentStep = Step.Otp;
+          this.resetForm.controls['email'].setValue(this.emailForm.value.email ?? null); // carry email to reset step
         },
         error: (err) => {
           this.errorMessage = err.error?.message || 'Failed to send reset email.';
           this.successMessage = '';
         }
       });
+  }
+
+  submitOtp() {
+    if (this.otpForm.invalid) return;
+    const resetCode = Object.values(this.otpForm.value).join('');
+    this._authApiService.VerifyCode({ resetCode })
+      .pipe(takeUntil(this.ngUnsubscribe))
+      .subscribe({
+        next: () => {
+          this.errorMessage = '';
+          this.currentStep = Step.Reset;
+        },
+        error: (err) => {
+          this.errorMessage = err.error?.message || 'Invalid code.';
+          this.successMessage = '';
+        }
+      });
+  }
+
+  submitReset() {
+    if (this.resetForm.invalid) return;
+    this._authApiService.resetpass({ email: this.resetForm.get('email')?.value || '', newPassword: this.resetForm.get('newPassword')?.value || '' })
+      .pipe(takeUntil(this.ngUnsubscribe))
+      .subscribe({
+        next: () => {
+          this.errorMessage = '';
+          this.currentStep = Step.Email;
+          this.emailForm.reset();
+          this.otpForm.reset();
+          this.resetForm.reset();
+          this.router.navigate(['/auth/login']);
+        },
+        error: (err) => {
+          this.errorMessage = err.error?.message || 'Failed to reset password.';
+        }
+      });
+  }
+
+  moveToNext(event: any, nextInput: number) {
+    const input = event.target;
+    if (input.value.length === 1 && nextInput < 6) {
+      this.otpInputs.toArray()[nextInput].nativeElement.focus();
+    }
+    if (event.inputType === 'deleteContentBackward' && nextInput > 0 && !input.value) {
+      this.otpInputs.toArray()[nextInput - 1].nativeElement.focus();
+    }
   }
 
   ngOnDestroy() {
