--- apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/otp-code/otp-code.component.scss
@@ -1,203 +1,0 @@
-.form-box {
-  margin: auto !important;
-  width: 100%;
-  border-radius: 50px;
-  border-width: 1px;
-  gap: 8px;
-  padding: 40px;
-
-  form {
-    flex-direction: column;
-    gap: 0;
-  }
-
-  input {
-    width: 100%;
-    margin-bottom: 14px;
-    padding: 8px 16px;
-    border: 1.5px solid rgba(255, 255, 255, 0.18);
-    border-radius: 20px;
-    background: rgba(255, 255, 255, 0.10);
-    color: white;
-    outline: none;
-    font-size: 15px;
-    font-weight: 400;
-    transition: border 0.2s;
-    box-shadow: none;
-    gap: 8px;
-
-    &::placeholder {
-      color: #bbb;
-      opacity: 1;
-    }
-
-    &:focus {
-      border: 1.5px solid #ff4100;
-      background: rgba(255, 255, 255, 0.16);
-    }
-  }
-
-  .btn {
-    width: 100%;
-    background: #ff4100;
-    padding: 15px;
-    border: none;
-    border-radius: 12px;
-    color: white;
-    font-weight: bold;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background 0.3s;
-    margin-top: 8px;
-    margin-bottom: 0;
-
-    &:hover {
-      background: #e83e00;
-    }
-  }
-
-  .switch {
-    margin-top: 18px;
-    font-size: 15px;
-    text-align: center;
-
-    a {
-      color: #ff4100;
-      text-decoration: none;
-      font-weight: 500;
-    }
-  }
-}
-
-.error-message {
-  background: #ff4100;
-  color: #fff;
-  padding: 12px 20px;
-  border-radius: 10px;
-  margin: 16px auto 0 auto;
-  text-align: center;
-  font-size: 16px;
-  font-weight: 600;
-  max-width: 80%;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  letter-spacing: 0.5px;
-}
-
-.success-message {
-  background: #22c55e;
-  color: #fff;
-  padding: 12px 20px;
-  border-radius: 10px;
-  margin: 16px auto 0 auto;
-  text-align: center;
-  font-size: 16px;
-  font-weight: 600;
-  max-width: 80%;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  letter-spacing: 0.5px;
-}
-
-.otp-container {
-  background: rgba(30, 30, 30, 0.65);
-  border-radius: 32px;
-  border: 1.5px solid rgba(255, 255, 255, 0.35);
-  box-shadow: 0 4px 32px rgba(0, 0, 0, 0.25);
-  padding: 48px 36px 36px 36px;
-  max-width: 400px;
-  margin: 40px auto;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.otp-title {
-  font-size: 2.5rem;
-  font-weight: 800;
-  color: #fff;
-  text-align: center;
-  margin-bottom: 12px;
-  letter-spacing: 1px;
-}
-
-.otp-subtitle {
-  color: #fff;
-  font-size: 1.2rem;
-  text-align: center;
-  margin-bottom: 32px;
-  font-weight: 400;
-}
-
-.otp-inputs {
-  display: flex;
-  justify-content: center;
-  gap: 14px;
-  margin-bottom: 24px;
-}
-
-.otp-box {
-  width: 40px;
-  height: 48px;
-  font-size: 2rem;
-  text-align: center;
-  border: none;
-  border-bottom: 3px solid #fff;
-  background: transparent;
-  color: #ff4100;
-  font-weight: 700;
-  outline: none;
-  margin: 0 2px;
-  border-radius: 0;
-  transition: border-color 0.2s;
-}
-.otp-box:focus {
-  border-bottom: 3px solid #ff4100;
-}
-
-.btn {
-  width: 100%;
-  background: #ff4100;
-  padding: 15px;
-  border: none;
-  border-radius: 24px;
-  color: white;
-  font-weight: bold;
-  font-size: 18px;
-  cursor: pointer;
-  transition: background 0.3s;
-  margin-top: 8px;
-  margin-bottom: 18px;
-}
-.btn:hover {
-  background: #e83e00;
-}
-
-.resend-section {
-  text-align: center;
-  color: #fff;
-  font-size: 15px;
-  margin-top: 8px;
-}
-.resend-link {
-  color: #ff4100;
-  text-decoration: none;
-  font-weight: 600;
-  margin-left: 4px;
-  cursor: pointer;
-}
-.resend-link:hover {
-  text-decoration: underline;
-}
-
-@media (max-width: 500px) {
-  .otp-container {
-    padding: 24px 8px 16px 8px;
-    max-width: 98vw;
-  }
-  .otp-box {
-    width: 32px;
-    font-size: 1.3rem;
-  }
-  .otp-title {
-    font-size: 1.5rem;
-  }
-}
