--- apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/otp-code/otp-code.component.spec.ts
@@ -1,28 +1,0 @@
-/* tslint:disable:no-unused-variable */
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-
-import { OtpCodeComponent } from './otp-code.component';
-
-describe('OtpCodeComponent', () => {
-  let component: OtpCodeComponent;
-  let fixture: ComponentFixture<OtpCodeComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ OtpCodeComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(OtpCodeComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
