--- apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/reset-pass/reset-pass.component.html
@@ -1,29 +1,0 @@
-<div>
-  <p class="text-2xl font-bold text-center m-2">Reset Password</p>
-</div>
-
-<div *ngIf="errorMessage" class="error-message">{{ errorMessage }}</div>
-<div *ngIf="successMessage" class="success-message">{{ successMessage }}</div>
-
-<div class="form-box">
-  <form [formGroup]="resetForm" (ngSubmit)="submit()">
-    <input
-      type="email"
-      placeholder="Email"
-      formControlName="email"
-      required
-    />
-    <input
-      type="password"
-      placeholder="New Password"
-      formControlName="newPassword"
-      required
-    />
-    <button type="submit" class="btn" [disabled]="resetForm.invalid">
-      Reset Password
-    </button>
-    <p class="switch mt-2">
-      Back to <a routerLink="/auth/login">Login</a>
-    </p>
-  </form>
-</div>
