--- apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/reset-pass/reset-pass.component.scss
@@ -1,98 +1,0 @@
-.form-box {
-  margin: auto !important;
-  width: 80%;
-  border-radius: 50px;
-  border-width: 1px;
-  gap: 8px;
-  padding: 40px;
-
-  form {
-    flex-direction: column;
-    gap: 0;
-  }
-
-  input {
-    width: 100%;
-    margin-bottom: 14px;
-    padding: 8px 16px;
-    border: 1.5px solid rgba(255, 255, 255, 0.18);
-    border-radius: 20px;
-    background: rgba(255, 255, 255, 0.10);
-    color: white;
-    outline: none;
-    font-size: 15px;
-    font-weight: 400;
-    transition: border 0.2s;
-    box-shadow: none;
-    gap: 8px;
-
-    &::placeholder {
-      color: #bbb;
-      opacity: 1;
-    }
-
-    &:focus {
-      border: 1.5px solid #ff4100;
-      background: rgba(255, 255, 255, 0.16);
-    }
-  }
-
-  .btn {
-    width: 100%;
-    background: #ff4100;
-    padding: 15px;
-    border: none;
-    border-radius: 12px;
-    color: white;
-    font-weight: bold;
-    font-size: 18px;
-    cursor: pointer;
-    transition: background 0.3s;
-    margin-top: 8px;
-    margin-bottom: 0;
-
-    &:hover {
-      background: #e83e00;
-    }
-  }
-
-  .switch {
-    margin-top: 18px;
-    font-size: 15px;
-    text-align: center;
-
-    a {
-      color: #ff4100;
-      text-decoration: none;
-      font-weight: 500;
-    }
-  }
-}
-
-.error-message {
-  background: #ff4100;
-  color: #fff;
-  padding: 12px 20px;
-  border-radius: 10px;
-  margin: 16px auto 0 auto;
-  text-align: center;
-  font-size: 16px;
-  font-weight: 600;
-  max-width: 80%;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  letter-spacing: 0.5px;
-}
-
-.success-message {
-  background: #22c55e;
-  color: #fff;
-  padding: 12px 20px;
-  border-radius: 10px;
-  margin: 16px auto 0 auto;
-  text-align: center;
-  font-size: 16px;
-  font-weight: 600;
-  max-width: 80%;
-  box-shadow: 0 2px 8px rgba(0,0,0,0.08);
-  letter-spacing: 0.5px;
-}
