--- apps/SuperFitnessApp/src/app/core/auth/pages/components/ForgetPass/reset-pass/reset-pass.component.spec.ts
@@ -1,28 +1,0 @@
-/* tslint:disable:no-unused-variable */
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-
-import { ResetPassComponent } from './reset-pass.component';
-
-describe('ResetPassComponent', () => {
-  let component: ResetPassComponent;
-  let fixture: ComponentFixture<ResetPassComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ResetPassComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ResetPassComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
